refactor(export): extraire la capture overlay macOS dans le helper MacOSExport
